Use sshd -T instead of directly reading the configuration files
